[MM-50026] - Move purchase modal shared component, SeatsCalculator to own folder
