+++ components/seats_calculator/consequences.tsx
