+++ components/seats_calculator/index.tsx
