+++ components/seats_calculator/seats_calculator.scss
